+++ meili_flutter/lib/meili_flutter.dart
@@ -1,9 +1,6 @@
 export 'src/meili.dart';
-
-export 'src/model/avail_params.dart';
-export 'src/model/booking_params.dart';
-export 'src/model/flow_enum.dart';
-export 'src/model/meili_params.dart';
-
 export 'src/widgets/connect.dart';
 export 'src/widgets/meili_view.dart';
+
+export 'package:meili_flutter_platform_interface/meili_flutter_platform_interface.dart'
+    show MeiliParams, AvailParams, AdditionalParams, BookingParams, FlowType, MeiliFlutterPlatform;