+++ meili_flutter_android/lib/meili_flutter_android.dart
@@ -1,1 +1,15 @@
+import 'package:flutter/services.dart';
+import 'package:meili_flutter_platform_interface/meili_flutter_platform_interface.dart';
 
+class MeiliFlutterAndroid extends MeiliFlutterPlatform {
+  static const _channel = MethodChannel('meili_flutter');
+
+  static void registerWith() {
+    MeiliFlutterPlatform.instance = MeiliFlutterAndroid();
+  }
+
+  @override
+  Future<void> openMeiliView(MeiliParams params) {
+    return _channel.invokeMethod('openMeiliViewController', params.toMap());
+  }
+}