+++ meili_flutter_ios/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 0.3.1-beta.2
+
+- Updated license to proprietary.
+
+## 0.3.1-beta.1
+
+- Implemented `MeiliFlutterIos` with `registerWith()` for federated plugin registration.
+- Unified method channel name to `meili_flutter`.
+- Added `dartPluginClass` to pubspec for automatic platform registration.
+
 ## [0.1.0+1] - 2024-08-01
 
 ### Added